[FIX] account_asset: first depreciation amount
When an asset is set as Prorata Temporis, with a one month period, the
first depreciation amount is computed on the basis of the remaining days
in the purchase year. This doesn't make sense for a monthly
depreciation.

This commit makes the distinction between a monthly and a yearly
depreciation. In the case of a monthly depreciation, the remaining days
in the purchase month are taken into account.

opw-690034

Backport of 5702ca1e8fd39d57931bb80b4cad7dfcc38fc5d7

The fiscal year start date should be taken into account when calculating
the depreciation amount in case of prorata temporis.

opw-678072

Backport of 3e2c583d453a90cce9ff70077b0dd28485ba52e6