[FIX] stock: _compute_package_info when children are packages
When we put a pack in a pack and try to validate the picking, we get
an error saying we are trying to create a pack operation without a
location id.

This is caused by the 'quant.package' location computation, when we
validate a picking, odoo will create the pack operations in the next
picking if there are any. And when it tries to create a pack operation
of a pack in pack, it cannot get the location id of the top level
package quant because in the compution of the location_id, it only looks
at quants having a package_id correspeding to the package. And there are
not any when the it's a package of package.

To solve this issue, we fixed the function '_compute_children_quant_ids'
that doesn't work either because of the same reasons. And then we use it
in function '_compute_package_info' to get the quants if there are any.

This bug has been introduced in rev: https://github.com/odoo/odoo/commit/bf5e52cbe0cc013cc54b5cef71a00df146f4ae7e