[FIX] point_of_sale -> stock: handling of lots/serial numbers
Before, the quantities on the pack operations were set also for the
pack operations that should have lots and had not, which is an inconsistency.
As at the end of the PoS, errors are not shown and not rollbacked, this
created really nasty situations.  With this code, it was never meant to
use lots in the PoS, but as the picking was not validated (but existed, because
not rollbacked) some people thought it as a wanted feature.

In the mean time, we had commit a787743dd51, but it will give
problems as it will not automatically create a backorder.  It is
better not to do anything when lots are involved somewhere in the picking
as then the user will see the open pickings in the PoS Orders picking type.
It is also important to not do anything when it is a move without a picking.

Needs to be forward ported to saas-11, where the file structure changed.

opw: 692045