[IMP] stock: add tooltips on push and procurement rules
Push and procurement rules are some difficult concepts to understand.

Add some tooltips :

Push Rules tooltip :
When a move is foreseen to a location, the push rule will automatically create a move to a next location after. This is mainly only needed when creating manual operations e.g. 2/3 step manual purchase order or 2/3 step finished product manual manufacturing order. In other cases, it is important to use pull rules where you know where you are going based on a demand.

Procurement Rules :
The demand represented by a procurement from e.g. a sale order, a reordering rule, another move, ... needs to be solved by applying a procurement rule. Depending on the action on the procurement rule, this triggers a purchase order, manufacturing order or another move. This way we create chains in the reverse order from the endpoint with the original demand to the starting point. That way, it is always known where we need to go and that is why they are preferred over push rules.