[IMP] web: Don't display aggregates in list views for columns with multi-currency records
Purpose:
On a list view (Ex: the invoices list view), if a column is a currency column (Ex: total amount), it makes no sense to display the aggregate (Ex: sum) if the records contains several currencies.

Spec:
Don't display the aggregate of a column if a least two records are in different currencies.