[FIX] website: correct some snippet structure to ease edition
Wrap the text content of s_button and s_share snippets so that it can
be edited easily (on chrome, the icons which are next to the next give
the user a hard time).

Note: this is not a full stable fix as it requires an update to be
deployed but the fix is not critical and may so be ignored without
breaking anything.