[IMP] project: add a wizard to merge tasks
This wizard allows the users to merge some task in a new task or an existing one

A user and a project are taken by default but may be configurable on the wizard.

The other tasks are archived. A message is posted on the tasks to report the target and a message is posted on the target to report the origin tasks.