[FIX] stock: recheck quants availbility when partially available
When a picking is partially available, and the company works in 'just in
time', they often need to re-check the availability of some products
just before sending them.

In 8.0, it was totally possible, to recheck the availabilty of products.

To fix this, we just add a button when the picking state is 'partially
availbale', which does exactly the same as 'Reserve', but is secondary
and contains the label 'Recheck availability'.