[FIX] web: lazy loading the same file multiple times with loadJS
Before this commit, if the loadJS method was called two times with the
same url, the associated script was effectively put in the DOM only one
time. However, the second call returned deferred was immediatly
resolved while the script (added by the first call) might not be fully
loaded yet.

Now, calling loadJS for a same url multiple times will always return
the same deferred, resolved when the associated script is fully loaded.