[FIX] base: writing on is_company field
The commercial_partner_id field on "account.invoice" is a stored related
field to commercial_partner_id on res.partner and this field is recomputed
when changing the is_company field (by the function _commercial_partner_store_triggers).
So when changing the "is_company" field, the commercial_partner_id is recomputed and
a write is made on each account.invoice linked to the commercial_partner_id.

To avoid access rights problem on account.invoice, the write on "is_company"
must be done in SUPERUSER.

opw:694687