[FIX] stock_account: _get_anglo_saxon_price_unit
Backport of 25d0d1411d98c7a5a009c7ab8d99a59f25948fc8

When getting the price from a account.move.line, the uom set
on the line must be taken into account.

Before the fix:

-Set anglo-saxon accounting
-Create two units of measure Box and Mot where 1Mot = 10Box
-Create a stockable product with Box as unit of reference and set with
real price and perpetual
-Create a SO with a Mot of this product
-Confirm the SO and create the invoice
-Validate the invoice

Bug:
In the journal of entries, only one box was counted for the debit and the credit
instead of 10.

NB: The function _get_anglo_saxon_price_unit must return the price unit of the
product in the unit of measure of the account.invoice.line.

opw:693075