[FIX] web_editor: let whole backend know when leaving fullscreen editor
An enterprise feature was in conflict with the fullscreen mode of the
editor (see commit https://github.com/odoo/enterprise/commit/125414b73e4d839fc44c445bb83591e84f4c855d which fixes it).
This commit forces this feature to be retriggered when the fullscreen
mode is left (as the feature is now put asleep during the
fullscreen mode).

This should obviously be improved to handle fullscreen mode with a more
robust way.