[FIX] mrp: blocking a work order and workcenter state
The issue is that if you block a workorder when you exceeded the
expected time, the workcenter is not blocked.

The workcenter state is computed according to its productivity lines.
The productivy lines are marked with a loss type type that define the
reason of the time loss (availability problem like missing equipment or
quality problem like broken equipment). There are also two magic loss
type that define if the work on the workcenter is done in the expected
time (fully productive time) or is done exceeding the expected time
(reduced speed). The unblock button appears if the workcenter is
blocked.

To compute the workcenter state, we have to look through the last
productivity line of the workcenter having no end_date. Before the
patch, we look for the last productivity line and it could return one
with an end_date while there was a productivity line without one and it
was this one that would define the workcenter state.

This is solved by immediately searching for a record without end date.
We also add some comment to the code to understand a little more what's
going on.